fix: python3 division conversion to int

+++ opensfm/context.py
@@ -36,7 +36,7 @@ def parallel_map(func, args, num_proc, max_batch_size=1):
         return list(map(func, args))
     else:
         with parallel_backend('loky', n_jobs=num_proc):
-            batch_size = max(1, len(args)/(num_proc*2))
+            batch_size = max(1, int(len(args)/(num_proc*2)))
             batch_size = min(batch_size, max_batch_size) if max_batch_size else batch_size
             return Parallel(batch_size=batch_size)(delayed(func)(arg) for arg in args)
